requirements.txt
Flask==2.1.2
Werkzeug==2.2.2
google-cloud-storage
pymysql
pathlib
datetime